+++ src/allanguage/alLangServerProxy.ts
@@ -286,7 +286,10 @@ export class ALLangServerProxy {
         return "";
     }
 
-    async getPageList(resourceUri: vscode.Uri | undefined) : Promise<string[]> {
+    /**
+     * @deprecated This method should not be used. Please use getPagesList method from ToolsLangServerClient class
+     */
+     async getPageList(resourceUri: vscode.Uri | undefined) : Promise<string[]> {
         let fileContent = "page 0 _symbolcache\n{\nprocedure t()\nvar\nf:page ;\nbegin\nend;\n}";
         let list = await this.getCompletionForSourceCode(resourceUri, "Loading list of pages.", fileContent,
             4, 7, 7, 1);
@@ -297,17 +300,9 @@ export class ALLangServerProxy {
         return [];
     }
 
-    async getReportList(resourceUri: vscode.Uri | undefined) : Promise<string[]> {
-        let fileContent = "codeunit 0 _symbolcache\n{\nprocedure t()\nvar\nf:report ;\nbegin\nend;\n}";
-        let list = await this.getCompletionForSourceCode(resourceUri, "Loading list of reports.", fileContent,
-            4, 9, 7, 1);
-
-        //process results
-        if (list)
-            return this.getSymbolLabels(list, vscode.CompletionItemKind.Reference);
-        return [];
-    }
-
+    /**
+     * @deprecated This method should not be used. Please use getTablesList method from ToolsLangServerClient class
+     */
     async getTableList(resourceUri: vscode.Uri | undefined) : Promise<string[]> {
         let fileContent = "codeunit 0 _symbolcache\n{\nprocedure t()\nvar\nf:record ;\nbegin\nend;\n}";
         let list = await this.getCompletionForSourceCode(resourceUri, "Loading list of tables.", fileContent,
@@ -319,7 +314,10 @@ export class ALLangServerProxy {
         return [];
     }
 
-    async getCodeunitList(resourceUri: vscode.Uri | undefined) : Promise<string[]> {
+    /**
+     * @deprecated This method should not be used. Please use getCodeunitsList method from ToolsLangServerClient class
+     */
+     async getCodeunitList(resourceUri: vscode.Uri | undefined) : Promise<string[]> {
         let fileContent = "codeunit 0 _symbolcache\n{\nprocedure t()\nvar\nf:codeunit ;\nbegin\nend;\n}";
         let list = await this.getCompletionForSourceCode(resourceUri, "Loading list of codeunits.", fileContent,
             4, 11, 7, 1);
@@ -330,7 +328,10 @@ export class ALLangServerProxy {
         return [];
     }
 
-    async getInterfaceList(resourceUri: vscode.Uri | undefined) : Promise<string[]> {
+    /**
+     * @deprecated This method should not be used. Please use getInterfacesList method from ToolsLangServerClient class
+     */
+     async getInterfaceList(resourceUri: vscode.Uri | undefined) : Promise<string[]> {
         let fileContent = "codeunit 0 _symbolcache implements  \n{\n}";
         let list = await this.getCompletionForSourceCode(resourceUri, "Loading list of interfaces.", fileContent,
             0, 36, 2, 1);